pyproject.toml
[project]
name = "subplayer"
version = "0.1.0"
description = "Кроссплатформенный офлайн-плеер с авто-субтитрами и переводом"
requires-python = ">=3.9"
dependencies = [
    "PySide6>=6.6.0",
    "python-mpv>=1.0.3",
    "argostranslate>=1.9.0",
    "pysrt>=1.1.2",
    "faster-whisper>=1.0.0",
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

